refactor: SuccessAddedModal 네이밍 수정
