+++ src/components/templates/SuccessAddedContent/SuccessAddedContent.styles.js
