+++ src/components/templates/index.js
@@ -1,5 +1,5 @@
-import SuccessAddedModal from './SuccessAddedModal/SuccessAddedModal';
+import SuccessAddedContent from './SuccessAddedContent/SuccessAddedContent';
 import GlobalNavbar from './GlobalNavbar/GlobalNavbar';
 import OrderContainer from './OrderContainer/OrderContainer';
 
-export { SuccessAddedModal, GlobalNavbar, OrderContainer };
+export { SuccessAddedContent, GlobalNavbar, OrderContainer };
